// frontend/src/api/generated/zod/semestres/semestres.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * SIFES API
 * API do Sistema do Instituto Federal de Esperantina — Diário de Aula
 * OpenAPI spec version: 1.0
 */
import * as zod from 'zod';


export const semestresControllerCreateBodyNomeRegExp = new RegExp('^\\d{4}\\/[12]$');


export const SemestresControllerCreateBody = zod.object({
  "nome": zod.string().regex(semestresControllerCreateBodyNomeRegExp),
  "dataInicio": zod.string(),
  "dataFim": zod.string()
})

export const SemestresControllerCreateResponse = zod.object({
  "id": zod.string(),
  "nome": zod.string(),
  "dataInicio": zod.iso.datetime({"offset":true}),
  "dataFim": zod.iso.datetime({"offset":true})
})

export const SemestresControllerFindAllResponseItem = zod.object({
  "id": zod.string(),
  "nome": zod.string(),
  "dataInicio": zod.iso.datetime({"offset":true}),
  "dataFim": zod.iso.datetime({"offset":true})
})
export const SemestresControllerFindAllResponse = zod.array(SemestresControllerFindAllResponseItem)

export const SemestresControllerFindOneParams = zod.object({
  "id": zod.string()
})

export const SemestresControllerFindOneResponse = zod.unknown()

export const SemestresControllerUpdateParams = zod.object({
  "id": zod.string()
})

export const semestresControllerUpdateBodyNomeRegExp = new RegExp('^\\d{4}\\/[12]$');


export const SemestresControllerUpdateBody = zod.object({
  "nome": zod.string().regex(semestresControllerUpdateBodyNomeRegExp).optional(),
  "dataInicio": zod.string().optional(),
  "dataFim": zod.string().optional()
})

export const SemestresControllerUpdateResponse = zod.unknown()

export const SemestresControllerRemoveParams = zod.object({
  "id": zod.string()
})

export const SemestresControllerRemoveResponse = zod.unknown()